Add check to see if VM has already been migrated
This avoids trying to migrate an already migrated VM: "failed migration: The operation is not supported on the object". This shows up in vSphere with the following error: "The operation is not supported on the object. Migrating VM vm-GUID to different network without changing it's host is not supported. Please use Reconfigure API to change VM's network."

The check should really come before we attempt to find the source VM, but that will require some redesign.